requirements.txt
PyYAML==5.1.2
requests==2.22.0
colorama==0.4.1

